requirements.txt
APScheduler==3.9.1
async-generator==1.10
attrs==22.1.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
boto3==1.21.45
botocore==1.24.45
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
cryptography==38.0.1
exceptiongroup==1.0.0rc9
Flask==2.1.2
greenlet==1.1.2
h11==0.14.0
idna==3.4
importlib-metadata==4.12.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.0
MarkupSafe==2.1.1
mysql-connector-python==8.0.29
numpy==1.22.3
outcome==1.2.0
packaging==21.3
pandas==1.4.2
pluggy==1.0.0
protobuf==3.20.1
py==1.11.0
pycparser==2.21
PyMySQL==1.0.2
pyOpenSSL==22.1.0
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.1.2
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
s3transfer==0.5.2
selenium==4.2.0
six==1.15.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.35
tomli==2.0.1
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.12
urllib3-secure-extra==0.1.0
webdriver-manager==3.8.4
Werkzeug==2.1.2
wsproto==1.2.0
zipp==3.8.0
flask_login
flask_paginate
flask_sqlalchemy